requirements.txt
openai-whisper
torch
torchaudio
transformers
accelerate
soundfile
pydub
audioop-lts
numpy
scipy
scikit-learn
python-dotenv
tqdm
pyannote.audio>=3.1
speechbrain
omegaconf
openai
pydantic
